Added an Accelerometer Axis Freeze enum for freezing an axis when calibrating. Added a bias X and Y for accelerometer calibration to account for the offset when the phone is laying flat on a level surface.
Changed the AccelerometerMonitor class from a static class to non-static. Re-organized the static and non-static properties of the class with appropriate code regions. Made the class inherit from IDisposable for handling when an instance of the class is garbage collected (un-register Accelerometer event handlers).

Moved separate utility functions to a dedicated Utility class.

Created an interface for handling AccelerometerCalibration changes. Updated SensorCalibrationViewModel and CarControlPage to implement the interface.

Updated the App.xaml.cs page to handle App Stop and Resume, stopping and starting the Accelerometer. Handling the un-registering and re-registering of the Accelerometer listener and event handlers.